refactor: detekt警告箇所の修正
