+++ core/di/src/main/java/com/example/flush/core/di/RepositoryModule.kt
@@ -1,14 +1,14 @@
 package com.example.flush.core.di
 
+import com.example.flush.core.repository.AuthRepository
+import com.example.flush.core.repository.EmotionAnalysisRepository
+import com.example.flush.core.repository.ThrowingItemRepository
+import com.example.flush.core.repository.UserRepository
 import com.example.flush.data.api.EmotionAnalysisApi
 import com.example.flush.data.repository.AuthRepositoryImpl
 import com.example.flush.data.repository.EmotionAnalysisRepositoryImpl
 import com.example.flush.data.repository.ThrowingItemRepositoryImpl
 import com.example.flush.data.repository.UserRepositoryImpl
-import com.example.flush.core.repository.AuthRepository
-import com.example.flush.core.repository.EmotionAnalysisRepository
-import com.example.flush.core.repository.ThrowingItemRepository
-import com.example.flush.core.repository.UserRepository
 import com.google.android.gms.auth.api.identity.SignInClient
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
