+++ core/domain/src/main/java/com/example/flush/core/domain/SignUpWithEmailUseCase.kt
@@ -6,12 +6,12 @@ import com.example.flush.core.model.User
 import com.example.flush.core.repository.AuthRepository
 import com.example.flush.core.repository.UserRepository
 import com.github.michaelbull.result.Err
+import com.github.michaelbull.result.Result
 import com.github.michaelbull.result.andThen
 import com.github.michaelbull.result.map
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
-import com.github.michaelbull.result.Result
 
 class SignUpWithEmailUseCase @Inject constructor(
     private val authRepository: AuthRepository,
@@ -28,7 +28,7 @@ class SignUpWithEmailUseCase @Inject constructor(
                             email = firebaseUser.email ?: "",
                             name = "名無し",
                             iconUrl = "https://firebasestorage.googleapis.com/v0/b/flush-de17e.firebasestorage.app/o/" +
-                                    "default_icon.png?alt=media&token=c77ba166-ca8a-4504-b70e-c4749c8f9cfc",
+                                "default_icon.png?alt=media&token=c77ba166-ca8a-4504-b70e-c4749c8f9cfc",
                         )
                         userRepository.createUser(user)
                     }
