+++ core/model/src/main/java/com/example/flush/core/model/EmotionAnalysisModel.kt
@@ -1,6 +1,5 @@
 package com.example.flush.core.model
 
-
 import kotlinx.serialization.Serializable
 
 @Serializable
