+++ core/repository/src/main/java/com/example/flush/core/repository/EmotionAnalysisRepository.kt
@@ -2,7 +2,6 @@ package com.example.flush.core.repository
 
 import com.example.flush.core.model.Emotion
 
-
 interface EmotionAnalysisRepository {
     suspend fun analyzeEmotion(text: String): Emotion
