+++ feature/auth_selection/src/main/java/com/example/flush/feature/auth_selection/AuthSelectionViewModel.kt
@@ -4,5 +4,5 @@ import com.example.flush.core.base.BaseViewModel
 
 class AuthSelectionViewModel :
     BaseViewModel<AuthSelectionUiState, AuthSelectionUiEvent, AuthSelectionUiEffect>(
-        AuthSelectionUiState()
+        AuthSelectionUiState(),
     )
