+++ feature/post/src/main/java/com/example/flush/feature/post/PostScreenDimensions.kt
@@ -1,5 +1,5 @@
 package com.example.flush.feature.post
 
-data object AnimationConfig {
+data object PostScreenDimensions {
     const val AnimationDuration = 5000
 }
